move: #12 요청/응답 DTO 분리
