+++ src/main/java/com/ijaes/jeogiyo/auth/controller/AuthController.java
@@ -1,8 +1,8 @@
 package com.ijaes.jeogiyo.auth.controller;
 
-import com.ijaes.jeogiyo.auth.dto.AuthResponse;
-import com.ijaes.jeogiyo.auth.dto.LoginRequest;
-import com.ijaes.jeogiyo.auth.dto.SignUpRequest;
+import com.ijaes.jeogiyo.auth.dto.response.AuthResponse;
+import com.ijaes.jeogiyo.auth.dto.request.LoginRequest;
+import com.ijaes.jeogiyo.auth.dto.request.SignUpRequest;
 import com.ijaes.jeogiyo.auth.service.AuthService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
