+++ src/main/java/com/ijaes/jeogiyo/auth/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.ijaes.jeogiyo.auth.dto;
+package com.ijaes.jeogiyo.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
