+++ src/main/java/com/ijaes/jeogiyo/auth/dto/request/SignUpRequest.java
@@ -1,6 +1,5 @@
-package com.ijaes.jeogiyo.auth.dto;
+package com.ijaes.jeogiyo.auth.dto.request;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
