+++ src/main/java/com/ijaes/jeogiyo/auth/dto/response/AuthResponse.java
@@ -1,4 +1,4 @@
-package com.ijaes.jeogiyo.auth.dto;
+package com.ijaes.jeogiyo.auth.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
