+++ src/main/java/com/ijaes/jeogiyo/auth/service/AuthService.java
@@ -1,8 +1,8 @@
 package com.ijaes.jeogiyo.auth.service;
 
-import com.ijaes.jeogiyo.auth.dto.AuthResponse;
-import com.ijaes.jeogiyo.auth.dto.LoginRequest;
-import com.ijaes.jeogiyo.auth.dto.SignUpRequest;
+import com.ijaes.jeogiyo.auth.dto.response.AuthResponse;
+import com.ijaes.jeogiyo.auth.dto.request.LoginRequest;
+import com.ijaes.jeogiyo.auth.dto.request.SignUpRequest;
 import com.ijaes.jeogiyo.auth.entity.TokenBlacklist;
 import com.ijaes.jeogiyo.auth.repository.TokenBlacklistRepository;
 import com.ijaes.jeogiyo.auth.security.JwtUtil;
