+++ src/main/java/com/ijaes/jeogiyo/user/dto/request/UpdateRoleRequest.java
@@ -1,4 +1,4 @@
-package com.ijaes.jeogiyo.user.dto;
+package com.ijaes.jeogiyo.user.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
