+++ src/main/java/com/ijaes/jeogiyo/user/service/UserService.java
@@ -3,7 +3,11 @@ package com.ijaes.jeogiyo.user.service;
 import com.ijaes.jeogiyo.auth.validator.SignUpValidator;
 import com.ijaes.jeogiyo.common.exception.CustomException;
 import com.ijaes.jeogiyo.common.exception.ErrorCode;
-import com.ijaes.jeogiyo.user.dto.*;
+import com.ijaes.jeogiyo.user.dto.request.UpdateAddressRequest;
+import com.ijaes.jeogiyo.user.dto.request.UpdatePasswordRequest;
+import com.ijaes.jeogiyo.user.dto.request.UpdatePhoneNumberRequest;
+import com.ijaes.jeogiyo.user.dto.response.UserInfoResponse;
+import com.ijaes.jeogiyo.user.dto.response.UserUpdateResponse;
 import com.ijaes.jeogiyo.user.entity.User;
 import com.ijaes.jeogiyo.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
